Implement edit cabin data

+++ src/services/apiCabins.js
@@ -25,37 +25,48 @@ export async function deleteCabin(cabinId) {
   return cabins;
 }
 
-export async function createCabin(cabin) {
-  const imageName = `${Math.random()}-${cabin.image.name}`.replaceAll('/', '');
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+export async function createEditCabin(cabin, id) {
+  // console.log(cabin, id);
+  const hasImagePath = cabin.image?.startsWith?.(supabaseUrl);
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...cabin, image: imagePath }])
-    .select();
+  const imageName = `${Math.random()}-${cabin.image.name}`.replaceAll('/', '');
+  const imagePath = hasImagePath
+    ? cabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  // 1. Create/edit cabin
+  let query = supabase.from('cabins');
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...cabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...cabin, image: imagePath }).eq('id', id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error('Cabin could not be created');
   }
 
-  // 2. Upload image
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, cabin.image, {
-      cacheControl: '3600',
-      upsert: false,
-    });
+  if (!hasImagePath) {
+    // 2. Upload image
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, cabin.image, {
+        cacheControl: '3600',
+        upsert: false,
+      });
 
-  // 3. Delete the cabin if there was an error uploading the image
-  if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id);
-    console.error(storageError);
-    throw new Error(
-      'Cabin image could not be uploaded and the cabin was not created'
-    );
+    // 3. Delete the cabin if there was an error uploading the image
+    if (storageError) {
+      await supabase.from('cabins').delete().eq('id', data.id);
+      console.error(storageError);
+      throw new Error(
+        'Cabin image could not be uploaded and the cabin was not created'
+      );
+    }
   }
-
   return data;
 }
